requirements.txt
Django==2.*
django-photologue==3.*
